new-module-imports: Extract utils..
..to `lib/utils/new-module`
