+++ lib/rules/new-module-imports.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const EMBER_NAMESPACES = ['inject.controller', 'inject.service'];
 const MAPPING = require('ember-rfc176-data');
+const { populateMessage } = require('../utils/new-module');
 
 const GLOBALS = MAPPING.reduce((memo, exportDefinition) => {
   if (exportDefinition.deprecated) {
@@ -33,6 +33,23 @@ module.exports = {
   },
 
   create(context) {
+    function reportNestedProperties(properties, parent) {
+      properties.forEach(item => {
+        const match = GLOBALS[`Ember.${parent}.${item.key.name}`];
+
+        populateMessage(
+          {
+            node: item,
+            key: item.key.name,
+            match,
+            parent,
+            type: item.type,
+          },
+          context
+        );
+      });
+    }
+
     return {
       VariableDeclarator(node) {
         // Filter out non-Ember variable declarations
@@ -49,13 +66,16 @@ module.exports = {
           } else {
             const key = item.key.name;
             const match = GLOBALS[`Ember.${key}`];
-            populateMessage({
-              node: item,
-              customKey: key !== item.value.name ? item.value.name : null,
-              key,
-              match,
-              type: item.type,
-            });
+            populateMessage(
+              {
+                node: item,
+                customKey: key !== item.value.name ? item.value.name : null,
+                key,
+                match,
+                type: item.type,
+              },
+              context
+            );
           }
         });
       },
@@ -84,7 +104,7 @@ module.exports = {
           const key = fullName.replace(/^Ember\./, '');
           const match = GLOBALS[fullName];
 
-          const reportedError = populateMessage({ node, fullName, key, match });
+          const reportedError = populateMessage({ node, fullName, key, match }, context);
 
           if (reportedError) {
             // exit the loop after the first match was found
@@ -93,63 +113,5 @@ module.exports = {
         }
       },
     };
-
-    function reportNestedProperties(properties, parent) {
-      properties.forEach(item => {
-        const match = GLOBALS[`Ember.${parent}.${item.key.name}`];
-
-        populateMessage({
-          node: item,
-          key: item.key.name,
-          match,
-          parent,
-          type: item.type,
-        });
-      });
-    }
-
-    function populateMessage(obj) {
-      // if a given global path does not exist in `globals.json` there is no
-      // JS module import for it, so do not report the error
-      if (!obj.match) {
-        return false;
-      }
-
-      // Both the Controller module and Service module each make available the
-      // "inject" namespace. In order to make it more readable, so it's more
-      // explicit at first glance from which module the "inject" namespace
-      // belongs to or is being imported from, let's check against this so we
-      // can properly populate the import specifier to report the ESLint error.
-      // Ex: import { inject as service } from '@ember/service';
-      const isNamespace = EMBER_NAMESPACES.indexOf(`${obj.parent}.${obj.key}`) !== -1;
-
-      const isNamedExport = obj.match.export !== 'default';
-
-      let importSpecifier;
-      let message;
-
-      if (isNamedExport && !isNamespace) {
-        importSpecifier = obj.match.localName
-          ? `{ ${obj.match.export} as ${obj.match.localName} }`
-          : `{ ${obj.match.export} }`;
-      } else if (isNamedExport && isNamespace) {
-        importSpecifier = `{ ${obj.parent} as ${obj.key} }`;
-      } else {
-        importSpecifier = obj.match.localName || obj.customKey || obj.key;
-      }
-
-      const replacement = `import ${importSpecifier} from '${obj.match.module}';`;
-
-      if (obj.type === 'Property') {
-        message = `Use ${replacement} instead of using Ember destructuring`;
-      } else {
-        message = `Use ${replacement} instead of using ${obj.fullName}`;
-      }
-
-      // report the issue to ESLint
-      context.report(obj.node, message);
-
-      return true;
-    }
   },
 };
